rds-manager/requirements.txt
django==1.10.5
djangorestframework
django-rest-swagger
ConfigParser
configparser
redis
rq
ruamel_yaml
pymysql
coreapi
django-sslserver
pytz
pyping
